certbotsh: Document reason for HTTP Digest authentication
It has very little impact on overall security (and thus disabled by
default) but might be useful in some environments to fight against of
certain attack types.

It has big side effect that in case of single host compromise password
must be changed on all the rest to keep this security mechanism working.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>